+++ ice-20201109/alibabacloud_ice20201109/client.py
@@ -533,6 +533,8 @@ class Client(OpenApiClient):
             query['SrtPassphrase'] = request.srt_passphrase
         if not DaraCore.is_null(request.srt_pbkey_len):
             query['SrtPbkeyLen'] = request.srt_pbkey_len
+        if not DaraCore.is_null(request.with_internal_vip):
+            query['WithInternalVip'] = request.with_internal_vip
         req = open_api_util_models.OpenApiRequest(
             query = Utils.query(query)
         )
@@ -581,6 +583,8 @@ class Client(OpenApiClient):
             query['SrtPassphrase'] = request.srt_passphrase
         if not DaraCore.is_null(request.srt_pbkey_len):
             query['SrtPbkeyLen'] = request.srt_pbkey_len
+        if not DaraCore.is_null(request.with_internal_vip):
+            query['WithInternalVip'] = request.with_internal_vip
         req = open_api_util_models.OpenApiRequest(
             query = Utils.query(query)
         )
@@ -4923,6 +4927,8 @@ class Client(OpenApiClient):
         query = {}
         if not DaraCore.is_null(request.file_ext):
             query['FileExt'] = request.file_ext
+        if not DaraCore.is_null(request.file_type):
+            query['FileType'] = request.file_type
         req = open_api_util_models.OpenApiRequest(
             query = Utils.query(query)
         )
@@ -4951,6 +4957,8 @@ class Client(OpenApiClient):
         query = {}
         if not DaraCore.is_null(request.file_ext):
             query['FileExt'] = request.file_ext
+        if not DaraCore.is_null(request.file_type):
+            query['FileType'] = request.file_type
         req = open_api_util_models.OpenApiRequest(
             query = Utils.query(query)
         )
@@ -12171,6 +12179,8 @@ class Client(OpenApiClient):
         query = {}
         if not DaraCore.is_null(request.flow_id):
             query['FlowId'] = request.flow_id
+        if not DaraCore.is_null(request.with_internal_vip):
+            query['WithInternalVip'] = request.with_internal_vip
         req = open_api_util_models.OpenApiRequest(
             query = Utils.query(query)
         )
@@ -12199,6 +12209,8 @@ class Client(OpenApiClient):
         query = {}
         if not DaraCore.is_null(request.flow_id):
             query['FlowId'] = request.flow_id
+        if not DaraCore.is_null(request.with_internal_vip):
+            query['WithInternalVip'] = request.with_internal_vip
         req = open_api_util_models.OpenApiRequest(
             query = Utils.query(query)
         )
@@ -14809,6 +14821,76 @@ class Client(OpenApiClient):
     ) -> main_models.GetYikeAssetMediaInfoResponse:
         runtime = RuntimeOptions()
         return await self.get_yike_asset_media_info_with_options_async(request, runtime)
+
+    def get_yike_storyboard_job_with_options(
+        self,
+        request: main_models.GetYikeStoryboardJobRequest,
+        runtime: RuntimeOptions,
+    ) -> main_models.GetYikeStoryboardJobResponse:
+        request.validate()
+        query = {}
+        if not DaraCore.is_null(request.job_id):
+            query['JobId'] = request.job_id
+        req = open_api_util_models.OpenApiRequest(
+            query = Utils.query(query)
+        )
+        params = open_api_util_models.Params(
+            action = 'GetYikeStoryboardJob',
+            version = '2020-11-09',
+            protocol = 'HTTPS',
+            pathname = '/',
+            method = 'POST',
+            auth_type = 'AK',
+            style = 'RPC',
+            req_body_type = 'formData',
+            body_type = 'json'
+        )
+        return DaraCore.from_map(
+            main_models.GetYikeStoryboardJobResponse(),
+            self.call_api(params, req, runtime)
+        )
+
+    async def get_yike_storyboard_job_with_options_async(
+        self,
+        request: main_models.GetYikeStoryboardJobRequest,
+        runtime: RuntimeOptions,
+    ) -> main_models.GetYikeStoryboardJobResponse:
+        request.validate()
+        query = {}
+        if not DaraCore.is_null(request.job_id):
+            query['JobId'] = request.job_id
+        req = open_api_util_models.OpenApiRequest(
+            query = Utils.query(query)
+        )
+        params = open_api_util_models.Params(
+            action = 'GetYikeStoryboardJob',
+            version = '2020-11-09',
+            protocol = 'HTTPS',
+            pathname = '/',
+            method = 'POST',
+            auth_type = 'AK',
+            style = 'RPC',
+            req_body_type = 'formData',
+            body_type = 'json'
+        )
+        return DaraCore.from_map(
+            main_models.GetYikeStoryboardJobResponse(),
+            await self.call_api_async(params, req, runtime)
+        )
+
+    def get_yike_storyboard_job(
+        self,
+        request: main_models.GetYikeStoryboardJobRequest,
+    ) -> main_models.GetYikeStoryboardJobResponse:
+        runtime = RuntimeOptions()
+        return self.get_yike_storyboard_job_with_options(request, runtime)
+
+    async def get_yike_storyboard_job_async(
+        self,
+        request: main_models.GetYikeStoryboardJobRequest,
+    ) -> main_models.GetYikeStoryboardJobResponse:
+        runtime = RuntimeOptions()
+        return await self.get_yike_storyboard_job_with_options_async(request, runtime)
 
     def insert_media_to_search_lib_with_options(
         self,
@@ -29083,6 +29165,124 @@ class Client(OpenApiClient):
     ) -> main_models.SubmitYikeAIAppJobResponse:
         runtime = RuntimeOptions()
         return await self.submit_yike_aiapp_job_with_options_async(request, runtime)
+
+    def submit_yike_storyboard_job_with_options(
+        self,
+        request: main_models.SubmitYikeStoryboardJobRequest,
+        runtime: RuntimeOptions,
+    ) -> main_models.SubmitYikeStoryboardJobResponse:
+        request.validate()
+        query = {}
+        if not DaraCore.is_null(request.aspect_ratio):
+            query['AspectRatio'] = request.aspect_ratio
+        if not DaraCore.is_null(request.model_params):
+            query['ModelParams'] = request.model_params
+        if not DaraCore.is_null(request.narration_voice_id):
+            query['NarrationVoiceId'] = request.narration_voice_id
+        if not DaraCore.is_null(request.resolution):
+            query['Resolution'] = request.resolution
+        if not DaraCore.is_null(request.shot_prompt_mode):
+            query['ShotPromptMode'] = request.shot_prompt_mode
+        if not DaraCore.is_null(request.title):
+            query['Title'] = request.title
+        if not DaraCore.is_null(request.user_data):
+            query['UserData'] = request.user_data
+        if not DaraCore.is_null(request.video_model):
+            query['VideoModel'] = request.video_model
+        body = {}
+        if not DaraCore.is_null(request.file_url):
+            body['FileURL'] = request.file_url
+        if not DaraCore.is_null(request.shot_split_mode):
+            body['ShotSplitMode'] = request.shot_split_mode
+        if not DaraCore.is_null(request.source_type):
+            body['SourceType'] = request.source_type
+        if not DaraCore.is_null(request.style_id):
+            body['StyleId'] = request.style_id
+        req = open_api_util_models.OpenApiRequest(
+            query = Utils.query(query),
+            body = Utils.parse_to_map(body)
+        )
+        params = open_api_util_models.Params(
+            action = 'SubmitYikeStoryboardJob',
+            version = '2020-11-09',
+            protocol = 'HTTPS',
+            pathname = '/',
+            method = 'POST',
+            auth_type = 'AK',
+            style = 'RPC',
+            req_body_type = 'formData',
+            body_type = 'json'
+        )
+        return DaraCore.from_map(
+            main_models.SubmitYikeStoryboardJobResponse(),
+            self.call_api(params, req, runtime)
+        )
+
+    async def submit_yike_storyboard_job_with_options_async(
+        self,
+        request: main_models.SubmitYikeStoryboardJobRequest,
+        runtime: RuntimeOptions,
+    ) -> main_models.SubmitYikeStoryboardJobResponse:
+        request.validate()
+        query = {}
+        if not DaraCore.is_null(request.aspect_ratio):
+            query['AspectRatio'] = request.aspect_ratio
+        if not DaraCore.is_null(request.model_params):
+            query['ModelParams'] = request.model_params
+        if not DaraCore.is_null(request.narration_voice_id):
+            query['NarrationVoiceId'] = request.narration_voice_id
+        if not DaraCore.is_null(request.resolution):
+            query['Resolution'] = request.resolution
+        if not DaraCore.is_null(request.shot_prompt_mode):
+            query['ShotPromptMode'] = request.shot_prompt_mode
+        if not DaraCore.is_null(request.title):
+            query['Title'] = request.title
+        if not DaraCore.is_null(request.user_data):
+            query['UserData'] = request.user_data
+        if not DaraCore.is_null(request.video_model):
+            query['VideoModel'] = request.video_model
+        body = {}
+        if not DaraCore.is_null(request.file_url):
+            body['FileURL'] = request.file_url
+        if not DaraCore.is_null(request.shot_split_mode):
+            body['ShotSplitMode'] = request.shot_split_mode
+        if not DaraCore.is_null(request.source_type):
+            body['SourceType'] = request.source_type
+        if not DaraCore.is_null(request.style_id):
+            body['StyleId'] = request.style_id
+        req = open_api_util_models.OpenApiRequest(
+            query = Utils.query(query),
+            body = Utils.parse_to_map(body)
+        )
+        params = open_api_util_models.Params(
+            action = 'SubmitYikeStoryboardJob',
+            version = '2020-11-09',
+            protocol = 'HTTPS',
+            pathname = '/',
+            method = 'POST',
+            auth_type = 'AK',
+            style = 'RPC',
+            req_body_type = 'formData',
+            body_type = 'json'
+        )
+        return DaraCore.from_map(
+            main_models.SubmitYikeStoryboardJobResponse(),
+            await self.call_api_async(params, req, runtime)
+        )
+
+    def submit_yike_storyboard_job(
+        self,
+        request: main_models.SubmitYikeStoryboardJobRequest,
+    ) -> main_models.SubmitYikeStoryboardJobResponse:
+        runtime = RuntimeOptions()
+        return self.submit_yike_storyboard_job_with_options(request, runtime)
+
+    async def submit_yike_storyboard_job_async(
+        self,
+        request: main_models.SubmitYikeStoryboardJobRequest,
+    ) -> main_models.SubmitYikeStoryboardJobResponse:
+        runtime = RuntimeOptions()
+        return await self.submit_yike_storyboard_job_with_options_async(request, runtime)
 
     def takeover_aiagent_call_with_options(
         self,