+++ ice-20201109/alibabacloud_ice20201109/models/_get_media_connect_flow_input_request.py
@@ -8,11 +8,13 @@ class GetMediaConnectFlowInputRequest(DaraModel):
     def __init__(
         self,
         flow_id: str = None,
+        with_internal_vip: str = None,
     ):
         # The flow ID.
         # 
         # This parameter is required.
         self.flow_id = flow_id
+        self.with_internal_vip = with_internal_vip
 
     def validate(self):
         pass
@@ -25,12 +27,18 @@ class GetMediaConnectFlowInputRequest(DaraModel):
         if self.flow_id is not None:
             result['FlowId'] = self.flow_id
 
+        if self.with_internal_vip is not None:
+            result['WithInternalVip'] = self.with_internal_vip
+
         return result
 
     def from_map(self, m: dict = None):
         m = m or dict()
         if m.get('FlowId') is not None:
             self.flow_id = m.get('FlowId')
+
+        if m.get('WithInternalVip') is not None:
+            self.with_internal_vip = m.get('WithInternalVip')
 
         return self